requirements.txt
pandas>=2.1,<2.3
numpy>=1.24,<2.1
yfinance>=0.2.40
requests>=2.31
feedparser>=6.0
pytrends>=4.9
praw>=7.7
arxiv>=2.1.0
scikit-learn>=1.3
matplotlib>=3.8
seaborn>=0.13
pyyaml>=6.0
joblib>=1.3
vaderSentiment>=3.3
transformers>=4.44
torch>=2.0; platform_system != 'Darwin' 
dotenv>=1.2.0
gdeltdoc>=1.12.0